fix(editor): Move auto-expose toggle to the correct MCP settings page
