+++ packages/frontend/editor-ui/src/features/ai/mcpAccess/SettingsMCPView.test.ts
@@ -3,7 +3,7 @@ import { createTestingPinia } from '@pinia/testing';
 import { waitFor, within } from '@testing-library/vue';
 import userEvent from '@testing-library/user-event';
 import { createComponentRenderer } from '@/__tests__/render';
-import { mockedStore, type MockedStore } from '@/__tests__/utils';
+import { mockedStore, type MockedStore, waitAllPromises } from '@/__tests__/utils';
 import SettingsMCPView from '@/features/ai/mcpAccess/SettingsMCPView.vue';
 import { useMCPStore } from '@/features/ai/mcpAccess/mcp.store';
 import { useSettingsStore } from '@n8n/stores/settings.store';
@@ -16,14 +16,22 @@ import { useExposeAllWorkflowsToMcpStore } from '@/experiments/exposeAllWorkflow
 import type { Agent } from '@/features/agents/agent.types';
 
 import { UNKNOWN_COUNT_VALUE } from '@/features/ai/mcpAccess/mcp.constants';
+import { TELEMETRY_EVENT } from '@n8n/telemetry';
 
 const { routerPush } = vi.hoisted(() => ({ routerPush: vi.fn() }));
 const { hasPermissionMock } = vi.hoisted(() => ({
 	hasPermissionMock: vi.fn().mockReturnValue(true),
 }));
+const { trackSpy } = vi.hoisted(() => ({
+	trackSpy: vi.fn(),
+}));
 
 vi.mock('@/app/utils/rbac/permissions', () => ({
 	hasPermission: hasPermissionMock,
+}));
+
+vi.mock('@n8n/composables/useTelemetry', () => ({
+	useTelemetry: () => ({ track: trackSpy }),
 }));
 
 vi.mock('vue-router', async (importOriginal) => ({
@@ -603,6 +611,76 @@ describe('SettingsMCPView', () => {
 			await nextTick();
 
 			expect(mcpStore.getInstanceClientStats).not.toHaveBeenCalled();
+		});
+	});
+
+	describe('auto-expose toggle', () => {
+		beforeEach(() => {
+			enableMcpSettings();
+		});
+
+		it('renders for a user with mcp:manage when the experiment is on', async () => {
+			hasPermissionMock.mockReturnValue(true);
+			exposeAllWorkflowsToMcpStore.isEnabled = true;
+
+			const { getByTestId } = createComponent({ pinia });
+			await waitAllPromises();
+
+			expect(getByTestId('mcp-auto-expose-toggle')).toBeTruthy();
+		});
+
+		it('is hidden without the experiment flag', async () => {
+			hasPermissionMock.mockReturnValue(true);
+			exposeAllWorkflowsToMcpStore.isEnabled = false;
+
+			const { queryByTestId } = createComponent({ pinia });
+			await waitAllPromises();
+
+			expect(queryByTestId('mcp-auto-expose-toggle')).toBeNull();
+		});
+
+		it('is hidden for a user without mcp:manage', async () => {
+			hasPermissionMock.mockReturnValue(false);
+			exposeAllWorkflowsToMcpStore.isEnabled = true;
+
+			const { queryByTestId } = createComponent({ pinia });
+			await waitAllPromises();
+
+			expect(queryByTestId('mcp-auto-expose-toggle')).toBeNull();
+		});
+
+		it('persists the new state and tracks the resulting value', async () => {
+			hasPermissionMock.mockReturnValue(true);
+			exposeAllWorkflowsToMcpStore.isEnabled = true;
+			mcpStore.setAutoExposeNewWorkflows.mockResolvedValue(true);
+
+			const { getByTestId } = createComponent({ pinia });
+			await waitAllPromises();
+
+			await userEvent.click(getByTestId('mcp-auto-expose-toggle').querySelector('input')!);
+
+			expect(mcpStore.setAutoExposeNewWorkflows).toHaveBeenCalledWith(true);
+			expect(trackSpy).toHaveBeenCalledWith(
+				TELEMETRY_EVENT.MCP.AUTO_EXPOSE_NEW_WORKFLOWS_TOGGLED,
+				expect.objectContaining({ enabled: true }),
+			);
+		});
+
+		it('shows a toast error and does not track when persisting fails', async () => {
+			hasPermissionMock.mockReturnValue(true);
+			exposeAllWorkflowsToMcpStore.isEnabled = true;
+			mcpStore.setAutoExposeNewWorkflows.mockRejectedValueOnce(new Error('nope'));
+
+			const { getByTestId } = createComponent({ pinia });
+			await waitAllPromises();
+
+			await userEvent.click(getByTestId('mcp-auto-expose-toggle').querySelector('input')!);
+			await waitAllPromises();
+
+			expect(trackSpy).not.toHaveBeenCalledWith(
+				TELEMETRY_EVENT.MCP.AUTO_EXPOSE_NEW_WORKFLOWS_TOGGLED,
+				expect.anything(),
+			);
 		});
 	});
 });
