+++ packages/frontend/editor-ui/src/features/ai/mcpAccess/SettingsMCPWorkflowsView.vue
@@ -2,12 +2,10 @@
 import { computed, onMounted, ref } from 'vue';
 import { useRouter } from 'vue-router';
 import { useI18n } from '@n8n/i18n';
-import { ElSwitch } from 'element-plus';
 import {
 	N8nButton,
 	N8nSettingsLayout,
 	N8nSettingsPageHeader,
-	N8nText,
 	N8nTooltip,
 } from '@n8n/design-system';
 import type { TableOptions } from '@n8n/design-system';
@@ -16,7 +14,6 @@ import { useDocumentTitle } from '@/app/composables/useDocumentTitle';
 import { useTelemetry } from '@n8n/composables/useTelemetry';
 import { useToast } from '@n8n/composables/useToast';
 import { useUIStore } from '@/app/stores/ui.store';
-import { hasPermission } from '@/app/utils/rbac/permissions';
 import { WORKFLOW_DESCRIPTION_MODAL_KEY } from '@/app/constants';
 import type { WorkflowListItem } from '@/Interface';
 import { useMCPStore } from '@/features/ai/mcpAccess/mcp.store';
@@ -27,10 +24,6 @@ import {
 	MCP_SETTINGS_VIEW,
 } from '@/features/ai/mcpAccess/mcp.constants';
 import WorkflowsTable from '@/features/ai/mcpAccess/components/tabs/WorkflowsTable.vue';
-import { useExposeAllWorkflowsToMcpStore } from '@/experiments/exposeAllWorkflowsToMcp/stores/exposeAllWorkflowsToMcp.store';
-import { EXPOSE_ALL_WORKFLOWS_TO_MCP_EXPERIMENT } from '@/app/constants/experiments';
-import { getExperimentTelemetryPayload } from '@/experiments/utils';
-import { TELEMETRY_EVENT } from '@n8n/telemetry';
 
 const i18n = useI18n();
 const toast = useToast();
@@ -39,7 +32,6 @@ const router = useRouter();
 const documentTitle = useDocumentTitle();
 const mcpStore = useMCPStore();
 const uiStore = useUIStore();
-const exposeAllWorkflowsToMcpStore = useExposeAllWorkflowsToMcpStore();
 
 const workflowsLoading = ref(false);
 const availableWorkflows = ref<WorkflowListItem[]>([]);
@@ -52,37 +44,6 @@ const workflowsTableState = ref<TableOptions>({
 const workflowsTableItemsPerPage = ref(workflowsTableState.value.itemsPerPage);
 
 const showConnectWorkflowsButton = computed(() => availableWorkflowsTotal.value > 0);
-
-const autoExposeSaving = ref(false);
-
-const canManageMcpInstance = computed(() =>
-	hasPermission(['rbac'], { rbac: { scope: 'mcp:manage' } }),
-);
-const showAutoExposeRow = computed(
-	() => canManageMcpInstance.value && exposeAllWorkflowsToMcpStore.isEnabled,
-);
-
-const onAutoExposeSwitchUpdate = (value: string | number | boolean) => {
-	void onToggleAutoExpose(value === true);
-};
-
-const onToggleAutoExpose = async (value: boolean) => {
-	autoExposeSaving.value = true;
-	try {
-		const updated = await mcpStore.setAutoExposeNewWorkflows(value);
-		telemetry.track(TELEMETRY_EVENT.MCP.AUTO_EXPOSE_NEW_WORKFLOWS_TOGGLED, {
-			enabled: updated,
-			...getExperimentTelemetryPayload(
-				EXPOSE_ALL_WORKFLOWS_TO_MCP_EXPERIMENT,
-				exposeAllWorkflowsToMcpStore.currentVariant,
-			),
-		});
-	} catch (error) {
-		toast.showError(error, i18n.baseText('settings.mcp.autoExpose.error.title'));
-	} finally {
-		autoExposeSaving.value = false;
-	}
-};
 
 const showMcpAccessUpdatedToast = (count: number, enabled: boolean) => {
 	toast.showMessage({
@@ -226,23 +187,6 @@ onMounted(async () => {
 			:docs-url="MCP_DOCS_PAGE_URL"
 		/>
 		<div data-test-id="mcp-workflows-view">
-			<div v-if="showAutoExposeRow" :class="$style.autoExposeRow">
-				<div :class="$style.autoExposeCopy">
-					<N8nText :bold="true" size="medium">
-						{{ i18n.baseText('settings.mcp.autoExpose.title') }}
-					</N8nText>
-					<N8nText size="small" color="text-light">
-						{{ i18n.baseText('settings.mcp.autoExpose.description') }}
-					</N8nText>
-				</div>
-				<ElSwitch
-					data-test-id="mcp-auto-expose-toggle"
-					:model-value="mcpStore.autoExposeNewWorkflows"
-					:disabled="mcpStore.mcpManagedByEnv"
-					:loading="autoExposeSaving"
-					@update:model-value="onAutoExposeSwitchUpdate"
-				/>
-			</div>
 			<div :class="$style.actions">
 				<N8nButton
 					v-if="showConnectWorkflowsButton"
@@ -299,19 +243,5 @@ onMounted(async () => {
 	justify-content: flex-end;
 	gap: var(--spacing--2xs);
 	margin-bottom: var(--spacing--xs);
-}
-
-.autoExposeRow {
-	display: flex;
-	align-items: center;
-	justify-content: space-between;
-	gap: var(--spacing--sm);
-	padding-bottom: var(--spacing--sm);
-}
-
-.autoExposeCopy {
-	display: flex;
-	flex-direction: column;
-	gap: var(--spacing--4xs);
 }
 </style>
